potholes-backend/potholes-backend/pyproject.toml
[project]
name = "potfoles"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.123.10",
    "google-generativeai>=0.8.5",
    "ipykernel>=7.1.0",
    "opencv-python-headless>=4.12.0.88",
    "pillow>=12.0.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "ultralytics>=8.3.235",
    "uvicorn>=0.38.0",
]
